// frontend/src/types/aas.ts
export enum ReferenceTypes {
  ExternalReference = 'ExternalReference',
  ModelReference = 'ModelReference',
}
// ...
export interface LangString {
  language: string;
  text: string;
}

export interface Key {
  type: string;
  value: string;
}

export interface Reference {
  type: ReferenceTypes;
  keys: Key[];
  referredSemanticId?: Reference;
}
// ...
export interface EmbeddedDataSpecification {
  dataSpecification: Reference;
  dataSpecificationContent: DataSpecificationIec61360;
}

export interface DataSpecificationIec61360 {
  preferredName: LangString[];
  shortName?: LangString[];
  unit?: string;
  unitId?: Reference;
  sourceOfDefinition?: string;
  symbol?: string;
  dataType?: string;
  definition?: LangString[];
  valueFormat?: string;
  valueList?: { valueReferencePairs: { value: string; valueId: Reference }[] };
  value?: string;
  levelType?: { min: boolean; nom: boolean; typ: boolean; max: boolean };
}